house_server/requirements.txt
asgiref==3.6.0
beautifulsoup4==4.11.2
category-encoders==2.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
dill==0.3.6
Django==4.1.7
djangorestframework==3.14.0
greenlet==2.0.2
idna==3.4
joblib==1.2.0
lxml==4.9.2
numpy==1.24.2
packaging==23.0
pandas==1.5.3
patsy==0.5.3
Pillow==9.4.0
psycopg2==2.9.5
psycopg2-binary==2.9.5
python-dateutil==2.8.2
python-decouple==3.7
pytz==2022.7.1
requests==2.28.2
scikit-learn==1.2.1
scipy==1.10.0
six==1.16.0
soupsieve==2.4
SQLAlchemy==2.0.3
sqlparse==0.4.3
statsmodels==0.13.5
threadpoolctl==3.1.0
typing_extensions==4.5.0
tzdata==2022.7
urllib3==1.26.14
xgboost==1.7.4
